Split into  vue component and data file
