+++ src/App.vue
@@ -4,18 +4,43 @@
       <component :is="Component"/>
     </keep-alive>
   </router-view>
-  <Footer></Footer>
+  <Footer :card-items="cardItems" :other-info-items="otherInfoItems" :icon-items="iconItems"></Footer>
 </template>
 
 <script lang="ts">
-import { defineComponent } from 'vue'
+import { defineComponent, ref } from 'vue'
 import Footer from '@/components/app/Footer.vue'
 import '@/assets/scss/app.scss'
+import footerData from '@/assets/json/footer-data.json'
 
 export default defineComponent({
   name: 'App',
   components: {
     Footer
+  },
+  setup() {
+    interface CardItem {
+      title: string;
+      content: string;
+      url: string;
+    }
+
+    interface OtherInfoItem {
+      title: string;
+      url: string;
+    }
+
+    interface IconItem {
+      name: string;
+      type: string;
+      url: string;
+    }
+
+    return {
+      cardItems: ref<Array<CardItem>>(footerData.cardItems),
+      otherInfoItems: ref<Array<OtherInfoItem>>(footerData.otherInfoItems),
+      iconItems: ref<Array<IconItem>>(footerData.iconItems)
+    }
   }
 })
 </script>
