+++ src/assets/json/footer-data.json
@@ -1,0 +1,46 @@
+{
+  "cardItems": [
+    {
+      "title": "辦公時間",
+      "content": "<div class='my-auto'><h2 class='text-lg mb-2'>平日</h2>8:30 AM ~ 12:00 AM<br />13:30 PM ~ 17:00 PM</div>",
+      "url": "https://i.imgur.com/kcClJic.webp"
+    },
+    {
+      "title": "通訊地址",
+      "content": "<div>台北市大安區忠孝東路三段一號，台灣，地球</div>",
+      "url": "https://i.imgur.com/h0fR1wm.webp"
+    },
+    {
+      "title": "聯絡我們",
+      "content": "<div>None</div>",
+      "url": "https://i.imgur.com/2yEBf8o.webp"
+    }
+  ],
+  "otherInfoItems": [
+    {
+      "title": "網站地圖",
+      "url": "#"
+    },
+    {
+      "title": "隱私權政策",
+      "url": "#"
+    }
+  ],
+  "iconItems": [
+    {
+      "name":"instagram",
+      "type": "fab",
+      "url": "#"
+    },
+    {
+      "name": "facebook",
+      "type": "fab",
+      "url": "#"
+    },
+    {
+      "name": "github",
+      "type": "fab",
+      "url": "#"
+    }
+  ]
+}
